Server now supports multiple maps and map switching.
This commits contains a lot of changes compared to the previous one.
Essentially all changes to support multiple maps and map switching are contained in this commit.

A few new functions where added :
  pk_innit: Innitialising of the server and maps.
  clean: Removes unwanted characters from strings("\n", " " and "\r" at the moment).
  get_map_list: Central process that knows about all maps.
  ch_map: permits to change actif map.

Warning: At the moment number of maps is hardecoded and is limited to two. This will be changed, it served as an experimentation.